Setup the foundation for future localization
I think this is the right call! This is fundementally an english project,
so the units of localization are done on the recipe level,
and I want discoverability to be maximized per language, so you can navigate to your language
directory and see at a glance the state of things. It does make the locale field a bit redudant
but not by much!